% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\name{apply_straying}
\alias{apply_straying}
\title{Apply Straying}
\usage{
apply_straying(
  year,
  natural_adults,
  hatchery_adults,
  total_releases,
  release_month,
  flows_oct_nov,
  flows_apr_may,
  monthly_mean_pdo
)
}
\arguments{
\item{year}{Simulation year}

\item{natural_adults}{Number of natural adults}

\item{hatchery_adults}{Number of hatchery adults}

\item{total_releases}{Total hatchery fish released from..params$hatchery_release,}

\item{release_month}{Hatchery release month, defaults to 1,}

\item{flows_oct_nov}{Mean flows from October and November from ..params$flows_oct_nov,}

\item{flows_apr_may}{Mean flows from April and May ..params$flows_apr_may,}

\item{monthly_mean_pdo}{Mean pdo from winterRunDSM::monthly_mean_pdo,}
}
\description{
Function to apply adult upstream straying
}
\details{
Apply Straying
}
