% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_list_function.R
\name{calculate_habitat_additions_ASD_BC}
\alias{calculate_habitat_additions_ASD_BC}
\title{Calculate habitat additions for ASD actions}
\usage{
calculate_habitat_additions_ASD_BC()
}
\description{
Calculates the values of added habitat across ASD actions
}
