% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\name{compute_adult_stray_rates}
\alias{compute_adult_stray_rates}
\title{Adult Straying for Hatchery Origin Fish}
\usage{
compute_adult_stray_rates(
  type = c("natural", "hatchery"),
  sim_year,
  total_releases,
  released_month,
  flows_oct_nov,
  flows_apr_may,
  mean_pdo_return
)
}
\arguments{
\item{mean_pdo_return}{PDO return}

\item{release_type}{river or bay representing where the hatchery fish were released}

\item{run_year}{year of run}

\item{age}{age of fish}

\item{released}{total number of fish released at hatchery}

\item{flow_oct_nov}{the median flow for October and November}

\item{flow_apr_may}{the median flow for April and May}
}
\description{
Calculates stray rates for all hatchery originating fish.
}
