% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_list_function.R
\name{convert_hab_matrix_to_df}
\alias{convert_hab_matrix_to_df}
\title{Convert DSMhabitat matrix to tidy df}
\usage{
convert_hab_matrix_to_df(matrix, scenario)
}
\description{
Converts a DSM habitat matrix to a tidy df to make plotting and comparison easier
}
\details{
input matrix should already be indexed for a watershed, giving a matrix of dim 12 x 20 
indicating month x sim year. produces results in acres with a column for month and for scenario.
}
