% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill_regional_dens_depend}
\alias{fill_regional_dens_depend}
\title{Fill Regional - Density Dependent Fill Hypothesis}
\usage{
fill_regional_dens_depend(
  juveniles,
  habitat,
  floodplain_habitat = NULL,
  up_to_size_class = 3,
  ..floodplain_capacity = 5,
  ..habitat_capacity = 4,
  ...
)
}
\arguments{
\item{..habitat_capacity}{the maximum number of rearing juveniles in channel habitat}

\item{floodplain_capacity}{the maximum number of rearing juveniles in floodplain habitat}
}
\description{
Alternative density dependent version of fill_regional with two additional parameters
}
