% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-baseline-data.R
\name{load_2019_baseline_data}
\alias{load_2019_baseline_data}
\title{Load 2019 Baseline Data}
\usage{
load_2019_baseline_data()
}
\description{
loads baseline data from OG model and calibrates for model comparisons
}
