% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natural_adult_removal_rate}
\alias{natural_adult_removal_rate}
\title{Spawners Removal Rate}
\format{
1 dimensional array [31 watersheds]
}
\source{
Spawner removal rate was derived using coded wire tag analysis. Details and methodology are available \href{https://cvpia-osc.github.io/fallRunDSM/articles/hatchery-analysis.html}{here}.

Compiled by: James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
natural_adult_removal_rate
}
\description{
Spawners removed for hatcheries
}
\details{
Spawners Removal Rate
}
\keyword{datasets}
