% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\name{normalize_with_context}
\alias{normalize_with_context}
\title{Normalize data with context data}
\usage{
normalize_with_context(x, context_data)
}
\description{
transform data to be normalized given data to calculate mean and standard deviation from
}
\keyword{internal}
