% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\name{normalize_with_params}
\alias{normalize_with_params}
\title{Normalize data with known params}
\usage{
normalize_with_params(x, mean_val, sd_val)
}
\description{
transform data to be normalized given the mean and standard deviation from the data
}
\keyword{internal}
