% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{prepare_stray_model_data}
\alias{prepare_stray_model_data}
\title{Prepare Beta-regression Data}
\usage{
prepare_stray_model_data(
  hatchery,
  type = c("natural", "hatchery"),
  sim_year,
  flow_oct_nov,
  flow_apr_may,
  releases,
  mean_PDO_return
)
}
\arguments{
\item{hatchery}{name of hatchery to produce data for}

\item{type}{either in river or bay release}

\item{sim_year}{simulation year, used to calculate run year}

\item{flow_oct_nov}{the mean flow for Oct-Nov, use DSMflow::hatchery_oct_nov_flows}

\item{flow_apr_may}{the mean flow for Apr-May, use DSMflow::hatchery_apr_may_flows}

\item{mean_PDO_return}{...}

\item{released}{the total number of released hatchery fish}
}
\description{
Create data frame of model data to be used in the beta-regression model for stray rates
}
\details{
Natural origin fish get a stray rate equal to Feather stray at distance = 0.
}
\keyword{internal}
