% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn_success.R
\name{spawn_success}
\alias{spawn_success}
\title{Spawning Success}
\source{
IP-117068
}
\usage{
spawn_success(
  escapement,
  proportion_natural,
  hatchery_age_distribution,
  natural_age_distribution,
  fecundity_lookup,
  adult_prespawn_survival,
  adult_prespawn_survival_abv_dam,
  abv_dam_spawn_proportion,
  abv_dam_spawn_habitat_proportion,
  dam_passage_survival,
  juvenile_capture_efficiency_dam_transport,
  harvest_rate_abv_dam,
  egg_to_fry_survival,
  egg_to_fry_survival_mult,
  egg_to_fry_survival_abv_dam,
  prob_scour,
  spawn_habitat,
  stochastic,
  sex_ratio,
  redd_size,
  fecundity
)
}
\arguments{
\item{escapement}{The number of returning adults}

\item{adult_prespawn_survival}{The adult prespawn survival rate}

\item{egg_to_fry_survival}{The egg to fry survival rate}

\item{prob_scour}{The probability of nest scouring}

\item{spawn_habitat}{The available spawning habitat in square meters}

\item{stochastic}{\code{TRUE} \code{FALSE} value indicating if model is being run stochastically}

\item{sex_ratio}{The female to male spawning ratio}

\item{redd_size}{The size of redds including defensible space}

\item{fecundity}{The number of eggs per female}
}
\description{
Calculates the annual reproductive success.
}
