% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stray.R
\name{stray_rates_to_matrix}
\alias{stray_rates_to_matrix}
\title{Creates matrix from stray rates}
\usage{
stray_rates_to_matrix(data, type)
}
\description{
transforms stray rates from betareg output into matrix
}
\keyword{internal}
