% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults.R
\name{surv_adult_enroute}
\alias{surv_adult_enroute}
\title{Adult En Route Survival}
\source{
IP-117068
}
\usage{
surv_adult_enroute(
  migratory_temp,
  bypass_overtopped,
  ..surv_adult_enroute_int =
    winterRunDSM::wr_sdm_baseline_params$..surv_adult_enroute_int,
  .migratory_temp = winterRunDSM::wr_sdm_baseline_params$.adult_en_route_migratory_temp,
  .bypass_overtopped =
    winterRunDSM::wr_sdm_baseline_params$.adult_en_route_bypass_overtopped,
  adult_enroute_surv_mult = winterRunDSM::wr_sdm_baseline_params$adult_enroute_surv_mult
)
}
\arguments{
\item{migratory_temp}{Variable representing proportion of migratory corridor temperature above  20°C}

\item{bypass_overtopped}{Indicator for bypass overtopped}

\item{..surv_adult_enroute_int}{Intercept}

\item{.migratory_temp}{Coefficient for \code{migratory_temp} variable}

\item{.bypass_overtopped}{Coefficient for \code{bypass_overtopped} variable}
}
\description{
Calculate adult survival en route to spawning grounds
}
\details{
See \code{\link{params}} for details on parameter sources
}
