% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_juv_outmigration_san_joaquin_flow_based}
\alias{surv_juv_outmigration_san_joaquin_flow_based}
\title{Juvenile Mainstem San Joaquin Outmigration Survival}
\source{
\link[DSMflow FF documentation]{https://reorienting-to-recovery.github.io/DSMflow/articles/EFF_SJ.html}
}
\usage{
surv_juv_outmigration_san_joaquin_flow_based(flow_cms)
}
\arguments{
\item{flow_cms}{Variable representing upper San Joaquin River flow in cubic meters per second}
}
\description{
Calculates the Mainstem San Joaquin juvenile out migration survival
}
