% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_list_function.R
\name{create_param_list}
\alias{create_param_list}
\title{Create parameter list for portfolios}
\usage{
create_param_list(action_ids)
}
\arguments{
\item{action_ids}{The action ids included in portfolio}
}
\description{
Creates updated parameter list for portfolios based on actions included
}
